requirements.txt
aiohttp==3.12.15
arabic_reshaper==3.0.0
beautifulsoup4==4.13.4
cairosvg==2.8.2
fitz==0.0.1.dev2
fonttools==4.59.0
groq==0.30.0
Markdown==3.8.2
matplotlib
numpy
opencv_python==4.12.0.88
pandas==2.3.1
Pillow==11.3.0
PyQt5==5.15.11
PyQt5_sip==12.17.0
pytesseract==0.3.13
python-dotenv==1.1.1
python_Levenshtein==0.27.1
Requests==2.32.4
tavily_python==0.7.10
tqdm==4.67.1
